+++ src/model/mod.rs
@@ -1,0 +1,9 @@
+mod building;
+mod employee;
+mod floor;
+mod seat;
+
+pub use self::building::Building;
+pub use self::employee::Employee;
+pub use self::floor::Floor;
+pub use self::seat::Seat;
